# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KG Extraction Backend",
    "description": "接收前端上传的 PDF，提取文本并保存为 TXT 文件",
    "version": "1.0.0"
  },
  "paths": {
...
    "/extract/run": {
      "post": {
        "tags": [
          "extract"
        ],
        "summary": "Run Graph Extraction",
        "operationId": "run_graph_extraction_extract_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/extract/run": {
      "post": {
        "tags": [
          "extract"
        ],
        "summary": "Api Run Graph Extraction",
        "operationId": "api_run_graph_extraction_api_extract_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/chunk/strategy": {
      "post": {
        "tags": [
          "chunk"
        ],
        "summary": "Create Split Strategy",
        "operationId": "create_split_strategy_chunk_strategy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitStrategyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chunk/strategy": {
      "post": {
        "tags": [
          "chunk"
        ],
        "summary": "Api Create Split Strategy",
        "operationId": "api_create_split_strategy_api_chunk_strategy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitStrategyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chunk/execute": {
      "post": {
        "tags": [
          "chunk"
        ],
        "summary": "Execute Chunking",
        "operationId": "execute_chunking_chunk_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chunk/execute": {
      "post": {
        "tags": [
          "chunk"
        ],
        "summary": "Api Execute Chunking",
        "operationId": "api_execute_chunking_api_chunk_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chunk/run": {
      "post": {
        "tags": [
          "chunk"
        ],
        "summary": "Run Strategy And Chunking",
        "operationId": "run_strategy_and_chunking_chunk_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chunk/run": {
      "post": {
        "tags": [
          "chunk"
        ],
        "summary": "Api Run Strategy And Chunking",
        "operationId": "api_run_strategy_and_chunking_api_chunk_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChunkExecutionRequest": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          }
        },
        "type": "object",
        "required": [
          "extraction_id"
        ],
        "title": "ChunkExecutionRequest"
      },
...
      "GraphExtractionRequest": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "auto"
          },
          "allowed_entity_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Entity Types"
          },
          "allowed_relation_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Relation Types"
          }
        },
        "type": "object",
        "required": [
          "extraction_id"
        ],
        "title": "GraphExtractionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SplitStrategyRequest": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          }
        },
        "type": "object",
        "required": [
          "extraction_id"
        ],
        "title": "SplitStrategyRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}